refactor(hermes): move rpc endpoints into submodules
