+++ hermes/src/api/rest/live.rs
@@ -1,0 +1,11 @@
+use axum::{
+    http::StatusCode,
+    response::{
+        IntoResponse,
+        Response,
+    },
+};
+
+pub async fn live() -> Response {
+    (StatusCode::OK, "OK").into_response()
+}
